src/main/webapp/app/entities/dm-loai-hd/delete/dm-loai-hd-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DmLoaiHdService } from '../service/dm-loai-hd.service';

import { DmLoaiHdDeleteDialogComponent } from './dm-loai-hd-delete-dialog.component';

describe('DmLoaiHd Management Delete Component', () => {
  let comp: DmLoaiHdDeleteDialogComponent;
  let fixture: ComponentFixture<DmLoaiHdDeleteDialogComponent>;
  let service: DmLoaiHdService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DmLoaiHdDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DmLoaiHdDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DmLoaiHdDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DmLoaiHdService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
